+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Release
 
+- Added the Grubbs Class (01/07/2022)
 - Added the Tukey Class (27/06/2022)
 - interquartile_range function was added (24/05/2021)
 - Dixon class was moved to outliers module (22/06/2022)