+++ docs/changelog.md
@@ -6,7 +6,8 @@
 
 ### Release
 
-- Added documentation for the Tukey tst for outlier (27/06/2022)
+- Added documentation for the Grubbstest for outlier detection (01/07/2022)
+- Added documentation for the Tukey test for outlier detection (27/06/2022)
 - interquartile_range documentation was added (24/05/2021)
 - Dixon class was moved to outliers module (22/06/2022)
 - Added documentation for the ModifiedZScore test (22/06/2022)